some changes

+++ server/src/controllers/noteController.ts
@@ -3,6 +3,7 @@ import express,{ Express,NextFunction,Request,RequestHandler,Response } from "ex
 import NoteModel from "../models/noteSchema"
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefine } from "../util/assertIsDefine";
 
 interface CreateNoteBody{
   title?:string;
@@ -11,9 +12,14 @@ interface CreateNoteBody{
 
 export const getNotes =  async(req: Request, res: Response, next:NextFunction) => {
 
+  const getAuthenticatedUserId = req.session.userId
+
     try {
       // throw createHttpError(401);
-        const notes= await NoteModel.find().exec();
+
+      assertIsDefine(getAuthenticatedUserId)
+
+        const notes= await NoteModel.find({userId: getAuthenticatedUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error)
@@ -24,12 +30,20 @@ export const getNotes =  async(req: Request, res: Response, next:NextFunction) =
   export const getNote =  async(req: Request, res: Response, next:NextFunction) => {
 
     const noteId= req.params.noteId;
+    const getAuthenticatedUserId = req.session.userId
+
     try {
+      assertIsDefine(getAuthenticatedUserId)
+
 
       if(!mongoose.isValidObjectId(noteId)) throw createHttpError(400,"invalid note id")
 
-      const newNotes=await NoteModel.findById(noteId).exec();
+      const newNotes =await NoteModel.findById(noteId).exec();
       if (!newNotes) throw createHttpError(404,"note not found");
+
+      if (newNotes && newNotes.userId && !newNotes.userId.equals(getAuthenticatedUserId)) {
+        throw createHttpError(401,"you cannot access this note")
+      }
 
       res.status(201).json(newNotes)
     } catch (error) {
@@ -42,9 +56,15 @@ export const getNotes =  async(req: Request, res: Response, next:NextFunction) =
   export const createNotes =  async(req: Request, res: Response, next:NextFunction) => {
 
     const {title,text}:CreateNoteBody = req.body ;
+    const getAuthenticatedUserId = req.session.userId
+
     try {
+      assertIsDefine(getAuthenticatedUserId)
+
+
       if(!title) throw createHttpError(400,"note must have a title")
       const newNotes=await NoteModel.create({
+    userId:getAuthenticatedUserId,
         title,
         text,
       })
@@ -58,7 +78,13 @@ export const getNotes =  async(req: Request, res: Response, next:NextFunction) =
     const noteId= req.params.noteId;
     const newTitle=req.body.title;
     const newText=req.body.text;
+
+    const getAuthenticatedUserId = req.session.userId
+
     try {
+
+      assertIsDefine(getAuthenticatedUserId)
+
 
       if(!mongoose.isValidObjectId(noteId)) throw createHttpError(400,"invalid note id")
       if(!newTitle) throw createHttpError(400,"note must have a title")
@@ -66,6 +92,11 @@ export const getNotes =  async(req: Request, res: Response, next:NextFunction) =
       const note=await NoteModel.findById(noteId).exec();
 
       if (!note) throw createHttpError(404,"note not found");
+
+      if (note && note.userId && !note.userId.equals(getAuthenticatedUserId)) {
+        throw createHttpError(401,"you cannot access this note")
+      }
+
 
       note.title= newTitle;
       note.text=newText;
@@ -82,12 +113,23 @@ export const getNotes =  async(req: Request, res: Response, next:NextFunction) =
 
     const noteId= req.params.noteId
 
+    const getAuthenticatedUserId = req.session.userId
+
+
     try {
+
+      assertIsDefine(getAuthenticatedUserId)
+
+
       if(!mongoose.isValidObjectId(noteId)) throw createHttpError(400,"invalid note id")
 
       const note= await NoteModel.findById(noteId).exec();
 
       if(!note) throw createHttpError(404,"note not found")
+
+      if (note && note.userId && !note.userId.equals(getAuthenticatedUserId)) {
+        throw createHttpError(401,"you cannot access this note")
+      }
 
       await note.deleteOne({_id:noteId});
       res.sendStatus(204);
